backend/requirements.txt

Flask==2.2.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.5
Werkzeug==2.2.3
PyJWT==2.6.0
python-dotenv==1.0.0
